refactor full_schedule
